feature/BA-0076 Resolved imports and classes references for web and rest modules
