--- rest/src/main/java/io/khasang/ba/config/WebAppInitializer.java
@@ -1,23 +1,0 @@
-package io.khasang.ba.config;
-
-import io.khasang.ba.config.application.WebConfig;
-import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
-
-// web.xml
-@Deprecated
-public class WebAppInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
-    @Override
-    protected Class<?>[] getRootConfigClasses() {
-        return new Class[]{WebConfig.class};
-    }
-
-    @Override
-    protected Class<?>[] getServletConfigClasses() {
-        return null;
-    }
-
-    @Override
-    protected String[] getServletMappings() {
-        return new String[]{"/"};
-    }
-}
