--- rest/src/main/java/io/khasang/ba/service/MyService.java
@@ -1,6 +1,0 @@
-package io.khasang.ba.service;
-
-public interface MyService {
-
-    String getName();
-}
