+++ web/src/main/java/io/khasang/ba/web/config/io/khasang/ba/web/config/WebAppInitializer.java
@@ -1,0 +1,21 @@
+package io.khasang.ba.web.config.io.khasang.ba.web.config;
+
+// web.xml
+@Deprecated
+public class WebAppInitializer {
+//        extends AbstractAnnotationConfigDispatcherServletInitializer {
+//    @Override
+//    protected Class<?>[] getRootConfigClasses() {
+//        return new Class[]{WebConfig.class};
+//    }
+//
+//    @Override
+//    protected Class<?>[] getServletConfigClasses() {
+//        return null;
+//    }
+//
+//    @Override
+//    protected String[] getServletMappings() {
+//        return new String[]{"/"};
+//    }
+}
