+++ web/src/main/java/io/khasang/ba/web/config/io/khasang/ba/web/config/application/FrontendApplication.java
@@ -1,0 +1,13 @@
+package io.khasang.ba.web.config.io.khasang.ba.web.config.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan({"io.khasang.ba.web"})
+public class FrontendApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(FrontendApplication.class, args);
+    }
+}
