+++ web/src/main/java/io/khasang/ba/web/config/io/khasang/ba/web/config/application/WebConfig.java
@@ -1,7 +1,6 @@
-package io.khasang.ba.config.application;
+package io.khasang.ba.web.config.io.khasang.ba.web.config.application;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
@@ -9,7 +8,6 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
 @Configuration
-@ComponentScan({"io.khasang.ba"})
 @EnableWebMvc
 public class WebConfig extends WebMvcConfigurerAdapter {
