+++ web/src/main/java/io/khasang/ba/web/config/io/khasang/ba/web/service/MyService.java
@@ -1,0 +1,6 @@
+package io.khasang.ba.web.config.io.khasang.ba.web.service;
+
+public interface MyService {
+
+    String getName();
+}
