+++ web/src/main/java/net/yandex/speller/services/spellservice/CheckTextsRequest.java
