+++ web/src/main/java/net/yandex/speller/services/spellservice/package-info.java
