+++ web/src/main/resources/log4j.properties
@@ -1,15 +1,13 @@
 # Root logger option
 log4j.rootLogger=INFO, stdout, file
-
 # redirect log messages to console
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
 log4j.appender.stdout.Target=System.out
 log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-
 # redirect log messages to file
 log4j.appender.file=org.apache.log4j.RollingFileAppender
-log4j.appender.file.File=c://temp//proj-back.log
+log4j.appender.file.File=c://temp//proj-front.log
 log4j.appender.file.MaxFileSize=5MB
 log4j.appender.file.MaxBackupIndex=10
 log4j.appender.file.layout=org.apache.log4j.PatternLayout
